+++ src/Dockwright/Data/Config.hs
@@ -1,30 +1,38 @@
-{-# LANGUAGE DataKinds     #-}
-{-# LANGUAGE TypeOperators #-}
-
-module Dockwright.Data.Config where
+module Dockwright.Data.Config
+  ( module X
+  , readConfig
+  , toValTagsConfig
+  , toRefTagsConfig
+  , splitOn
+  ) where
 
 import           RIO
+import qualified RIO.Text                        as Text
 
 import           Data.Extensible
+import qualified Data.Yaml                       as Y
+import           Dockwright.Data.Config.Default  as X
+import           Dockwright.Data.Config.Internal as X
 
-type Config = Record
-   '[ "output"   >: FilePath
-    , "template" >: FilePath
-    , "base"     >: BaseImageConfig
-    , "env"      >: Map String DockVal
-    ]
+readConfig :: MonadIO m => FilePath -> m (Either Y.ParseException Config)
+readConfig = readConfigWith defaultConfig
 
-type BaseImageConfig = Record
-   '[ "repo" >: Text
-    , "tag"  >: Text
-    ]
+readConfigWith ::
+  MonadIO m => Config -> FilePath -> m (Either Y.ParseException Config)
+readConfigWith def path = do
+  file <- readFileBinary path
+  pure $ case Y.decodeEither' file of
+    Right Y.Null -> Right def
+    _            -> hzipWith fromNullable def <$> Y.decodeEither' file
 
-type DockVal = Record
-   '[ "github" >: Maybe GitHubConfig
-    , "value"  >: Maybe Text
-    ]
+toValTagsConfig :: TagsConfig -> ValTagsConfig
+toValTagsConfig = shrink
 
-type GitHubConfig = Record
-   '[ "repo" >: Text
-    , "hook" >: Text
-    ]
+toRefTagsConfig :: TagsConfig -> RefTagsConfig
+toRefTagsConfig conf
+    = #ref @= fromMaybe "" (conf ^. #ref)
+   <: #keys @= (conf ^. #keys)
+   <: nil
+
+splitOn :: Char -> Text -> (Text, Text)
+splitOn c = second (Text.drop 1) . Text.span (/= c)