+++ src/Dockwright/Fetch.hs
@@ -1,46 +1,34 @@
-{-# LANGUAGE DataKinds        #-}
-{-# LANGUAGE LambdaCase       #-}
-{-# LANGUAGE OverloadedLabels #-}
-{-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeOperators    #-}
-
 module Dockwright.Fetch
   ( fetchEnvVal
   , Fetch (..)
   ) where
 
 import           RIO
-import qualified RIO.Text                as Text
 
 import           Data.Extensible
-import           Data.Proxy
 import           Dockwright.Data.Config
 import           Dockwright.Data.Env
 import           Dockwright.Fetch.GitHub
+import           Mix.Plugin.Logger       ()
 
-fetchEnvVal :: String -> DockVal -> RIO Env String
-fetchEnvVal key val =
-  maybe (throwM err) pure =<<
-    hfoldrWithIndexFor
-      (Proxy @ Fetch)
-      (\m v r -> liftA2 (<|>) (fetch m $ v ^. _Wrapper) r)
-      (pure Nothing)
-      val
-  where
-    err = FetchEnvError (Text.pack $ "there is no config to fetch env: " <> key)
+fetchEnvVal :: DockVal -> RIO Env (Either FetchError Text)
+fetchEnvVal val = fromMaybe (Left UndefinedConfig) <$>
+  hfoldrWithIndexFor
+    (Proxy @ Fetch)
+    (\m v r -> liftA2 (<|>) (fetch m $ v ^. _Wrapper) r)
+    (pure Nothing)
+    val
 
 class Fetch kv where
   fetch ::
-    (MonadIO m, MonadReader env m, HasLogFunc env)
-    => proxy kv -> AssocValue kv -> m (Maybe String)
+    (MonadUnliftIO m, MonadReader e m, HasLogFunc e)
+    => proxy kv -> TargetOf kv -> m (Maybe (Either FetchError Text))
 
 instance Fetch ("github" >: Maybe GitHubConfig) where
-  fetch _ = \case
-    Nothing   -> pure Nothing
-    Just conf -> case conf ^. #hook of
-      "release" -> fetchRelease' conf
-      key       ->
-        logWarn (display $ "unknown GitHub hook key:" <> key) >> pure Nothing
+  fetch _ Nothing = pure Nothing
+  fetch _ (Just conf) = case conf ^. #hook of
+    "release" -> Just <$> fetchRelease' conf
+    key       -> pure $ Just (Left $ UndefinedKey key)
 
 instance Fetch ("value" >: Maybe Text) where
-  fetch _ = pure . fmap Text.unpack
+  fetch _ = pure . fmap Right