+++ src/Dockwright/Fetch/GitHub.hs
@@ -1,41 +1,37 @@
-{-# LANGUAGE LambdaCase       #-}
-{-# LANGUAGE OverloadedLabels #-}
-
 module Dockwright.Fetch.GitHub where
 
 import           RIO
 import qualified RIO.Text               as Text
 
-import           Data.Default
-import           Dockwright.Data.Config (GitHubConfig)
-import           Dockwright.Data.Env    (DockwrightException (..))
+import           Data.Extensible
+import           Dockwright.Data.Config (GitHubConfig, splitOn)
+import           Dockwright.Data.Env
 import           Dockwright.Data.GitHub
+import           Mix.Plugin.Logger.JSON as MixLogger
 import           Network.HTTP.Req
 
 fetchRelease' ::
-  (MonadIO m, MonadReader e m, HasLogFunc e) => GitHubConfig -> m (Maybe String)
+  (MonadUnliftIO m, MonadReader e m, HasLogFunc e)
+  => GitHubConfig -> m (Either FetchError Text)
 fetchRelease' conf =
-  fetchRelease (splitOn '/' $ conf ^. #repo) >>= \case
-    Left err -> do
-      logWarn $ displayShow err
-      pure Nothing
-    Right r  -> pure $ Just (Text.unpack $ r ^. #tag_name)
+  fmap (stripByConfig conf . view #tag_name) <$> fetchRelease (splitOn '/' $ conf ^. #repo)
 
 fetchRelease ::
-  (MonadIO m, MonadReader e m, HasLogFunc e)
-  => (Text, Text) -> m (Either DockwrightException Release)
+  (MonadUnliftIO m, MonadReader e m, HasLogFunc e)
+  => (Text, Text) -> m (Either FetchError Release)
 fetchRelease (owner, repo) = do
-  logDebug (display $ "fetch github: " <> tshow url)
-  liftIO $ (latest <$> runReq def (responseBody <$> request)) `catch` handler
+  MixLogger.logDebugR "fetch release from github" (#url @= tshow url <: nil)
+  (latest <$> runReq defaultHttpConfig (responseBody <$> request)) `catch` handler
   where
     request = req GET url NoReqBody jsonResponse h
     url = https "api.github.com" /: "repos" /: owner /: repo /: "releases"
     h = header "User-Agent" "Dockwright"
-    handler :: MonadUnliftIO m => HttpException -> m (Either DockwrightException a)
-    handler = pure . Left . FetchEnvError . tshow
+    handler = pure . Left . HttpErr
 
-latest :: [Release] -> Either DockwrightException Release
-latest = maybe (Left $ FetchEnvError "no release") pure . listToMaybe
+latest :: [Release] -> Either FetchError Release
+latest = maybe (Left NoRelease) pure . listToMaybe
 
-splitOn :: Char -> Text -> (Text, Text)
-splitOn c = second (Text.drop 1) . Text.span (/= c)
+stripByConfig :: GitHubConfig -> Text -> Text
+stripByConfig conf release = fromMaybe release $ do
+  prefix <- conf ^. #strip_prefix
+  Text.stripPrefix prefix release