+++ test/Test.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE OverloadedLabels #-}
-
 module Test where
 
 import           RIO
@@ -11,15 +9,23 @@ import           Dockwright.Data.Env
 
 testConfig :: Config
 testConfig
-    = #output   @= "."
-   <: #template @= "./Dockerfile.template"
+    = #image    @= "matsubara0507/dockwright"
+   <: #output   @= "."
+   <: #template @= template
    <: #base     @= (#repo @= "debian" <: #tag @= "latest" <: nil)
-   <: #env      @= Map.fromList [("test1", testDockVal)]
+   <: #env      @= Just (Map.fromList [("test1", testDockVal)])
+   <: #tags     @= []
+   <: nil
+
+template :: DockerfileTeamplate
+template
+    = #before_env @= Nothing
+   <: #after_env  @= Just "./Dockerfile.template"
    <: nil
 
 testDockVal :: DockVal
 testDockVal
-    = #github @= Just (#repo @= "hoge/fuga" <: #hook @= "release" <: nil)
+    = #github @= Just (#repo @= "hoge/fuga" <: #hook @= "release" <: #strip_prefix @= Nothing <: nil)
    <: #value  @= Just "1.0.0"
    <: nil