+++ test/Test/Dockwright/Build.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE OverloadedLabels #-}
-
 module Test.Dockwright.Build where
 
 import           RIO
@@ -10,8 +8,8 @@ import           Language.Docker
 import           Test.Tasty
 import           Test.Tasty.Hspec
 
-test_buildBaseImage :: IO TestTree
-test_buildBaseImage = testSpec "buildBaseImage :: RIO Env Dockerfile" $
+tests :: IO TestTree
+tests = testSpec "buildBaseImage :: RIO Env Dockerfile" $
   context "when Env is testEnv" $
     it "correct case" $
       runRIO testEnv buildBaseImage `shouldReturn` toDockerfile (from $ "debian" `tagged` "latest")