+++ CHANGELOG.md
@@ -1,3 +1,20 @@
 # Changelog for dockwright
 
 ## Unreleased changes
+
+## 1.0.0
+
+- Refactor: update to lts14.4 and Stack v2
+    - update GHC to 8.6.5
+    - update extensible to 0.6.1
+    - update req to 2.1.0
+    - update language-docker to 8.0.2
+    - change package to githash from gitrev
+- Refactor: use mix.hs and fallible
+- Feat: strip prefix for release
+- Feat: before/after template
+- Modify: `--echo` can use lower case
+- Feat: `--tags` command
+- Feat: `--new-tags` command
+- Feat: generate default config
+- Feat: `--help` option